requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
textblob==0.17.1
reportlab==4.0.7
pydantic==2.5.0
openai==0.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
googletrans==4.0.0rc1
langdetect==1.0.9
